www/requirements.txt
Flask>=3.0.0
gunicorn>=21.2.0
psutil>=5.9.5
PyYAML>=6.0
apprise>=1.4.0
schedule>=1.2.0
confuse>=2.0.0
